moved cleargraph dependents to work with latest version 3.0.2
