+++ src/harmony/index.ts
@@ -1,3 +1,4 @@
+export { AnyExtension } from './any-extension';
 export { Extension } from './extension';
 export { default as Harmony } from './harmony';
 export { ProviderFn } from './types';
